# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Raudah Travels & Tours API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create notification (admin/system)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationBody"
      responses:
        "201":
          description: Notification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /support/tickets:
    get:
      operationId: listSupportTickets
      tags: [support]
      summary: List support tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved]
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Support tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketListResponse"
    post:
      operationId: createSupportTicket
      tags: [support]
      summary: Create support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportTicketBody"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /support/tickets/{id}:
    get:
      operationId: getSupportTicket
      tags: [support]
      summary: Get support ticket with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketDetail"
    put:
      operationId: updateSupportTicket
      tags: [support]
      summary: Update ticket status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupportTicketBody"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /support/tickets/{id}/messages:
    post:
      operationId: sendSupportMessage
      tags: [support]
      summary: Send a message in a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessage"

  # --- Public agent application (no auth) ---
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkUserId:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, agent, staff, admin, super_admin]
        ninNumber:
          type: string
          nullable: true
        passportNumber:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, clerkUserId, email, fullName, role, createdAt]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        type:
          type: string
          enum: [booking, payment, document, support, system]
        createdAt:
          type: string
      required: [id, userId, title, message, isRead, type, createdAt]
    CreateNotificationBody:
      type: object
      properties:
        userId:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [booking, payment, document, support, system]
      required: [userId, title, message, type]
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: number
      required: [notifications, unreadCount]
    SupportTicket:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        subject:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved]
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
        updatedAt:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
          nullable: true
      required: [id, userId, subject, status, priority, createdAt, updatedAt]
    SupportTicketDetail:
      allOf:
        - $ref: "#/components/schemas/SupportTicket"
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/SupportMessage"
          required: [messages]
    CreateSupportTicketBody:
      type: object
      properties:
        subject:
          type: string
        message:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        category:
          type: string
      required: [subject, message]
    UpdateSupportTicketBody:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved]
        priority:
          type: string
          enum: [low, medium, high]
    SupportMessage:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        senderId:
          type: string
        message:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
        sender:
          $ref: "#/components/schemas/UserProfile"
          nullable: true
      required: [id, ticketId, senderId, message, isAdmin, createdAt]
    CreateSupportMessageBody:
      type: object
      properties:
        message:
          type: string
      required: [message]
    SupportTicketListResponse:
      type: object
      properties:
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/SupportTicket"
        total:
          type: number
      required: [tickets, total]
...